Add "rolling colout"
Just realized one of the ways the old set_static_colour worked, was to accept a list of colours and create a movie that switchens between the colours.
Added that functionality back, but with the name "set_rolling_colour".  Please suggest a better name if you don't like it.

But that means that this is a kind of breaking change.